+++ lucentflow-common/src/main/java/com/lucentflow/common/repository/WhaleTransactionRepository.java
@@ -1,4 +1,4 @@
-package com.lucentflow.indexer.repository;
+package com.lucentflow.common.repository;
 
 import com.lucentflow.common.entity.WhaleTransaction;
 import org.springframework.data.domain.Page;