+++ lucentflow-indexer/src/main/java/com/lucentflow/indexer/sink/WhaleDatabaseSink.java
@@ -3,7 +3,7 @@ package com.lucentflow.indexer.sink;
 import com.lucentflow.common.entity.WhaleTransaction;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.lucentflow.indexer.repository.WhaleTransactionRepository;
+import com.lucentflow.common.repository.WhaleTransactionRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.jdbc.core.JdbcTemplate;